+++ playground/projects/sourcemap-minify/src/compute.js
@@ -1,0 +1,1 @@
+export const computeTheAnswer = (n) => n * 2;