+++ playground/projects/sourcemap-minify/src/wrapped.js
@@ -1,0 +1,3 @@
+import { computeTheAnswer } from './compute.js';
+
+export default computeTheAnswer(21);